Refactor component structure and update dependencies for improved organization and performance. Removed unused components and updated import paths for better clarity. Adjusted sitemap and package files to reflect recent changes in project structure.
