+++ package.json
@@ -13,8 +13,6 @@
     "@mdx-js/react": "^3.1.0",
     "@next/mdx": "^15.3.5",
     "@radix-ui/react-navigation-menu": "^1.2.5",
-    "@react-three/drei": "^10.5.0",
-    "@react-three/fiber": "^9.2.0",
     "clsx": "^2.1.1",
     "gray-matter": "^4.0.3",
     "highlight.js": "^11.11.1",
@@ -26,10 +24,8 @@
     "posthog-node": "^5.1.1",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
-    "react-tweet": "^3.2.2",
     "rehype-highlight": "^7.0.2",
-    "tailwind-merge": "^3.3.1",
-    "three": "^0.178.0"
+    "tailwind-merge": "^3.3.1"
   },
   "devDependencies": {
     "postcss": "^8",
