+++ src/app/coming-soon/page.js
@@ -1,8 +1,8 @@
 'use client';
 
 import { useEffect, useState } from 'react';
-import Navbar from '@/components/Navbar';
-import Footer from '@/components/NewFooter';
+import Navbar from '@/components/layout/Navbar';
+import Footer from '@/components/layout/Footer';
 
 export default function ComingSoonPage() {
   const [timeLeft, setTimeLeft] = useState({
