+++ src/components/sections/CaseStudies.jsx
