+++ src/components/sections/PricingSection.jsx
